fix(registry): add input validation + path traversal protection (P0.28)
- Package name regex validation (lowercase, alphanumeric + hyphen, 2+ chars)
- Path traversal protection in version strings
- Semver validation for version field
- Null byte injection checks
- 50MB request size limit via DefaultBodyLimit
- InvalidInput(String) error variant
- 22 new validation tests (all passing)
- Fix pre-existing clippy warnings in oauth.rs and test files

+++ crates/buff-registry/tests/allowlist_tests.rs
@@ -81,7 +81,7 @@ fn session_from_set_cookie(headers: &axum::http::HeaderMap) -> String {
     let prefix = "buff_session=";
     let start = cookie
         .find(prefix)
-        .expect("buff_session cookie") as usize
+        .expect("buff_session cookie")
         + prefix.len();
     let rest = &cookie[start..];
     let end = rest.find(';').unwrap_or(rest.len());
